Prevent a crash if not using a heat generated private key
It doesn't _seem_ valid that a private_key might come back from
heat as an empty string, but just in case that is valid, this
adds another check for None, which is what private_key will
be if the heat template does not output a private_key at all.

Without this, heat templates that do not generate and output
a private key will cause a crash, because None would be written
to the key file, and Python refuses to do that.